+++ game_patch/purefaction/pf.h
@@ -7,4 +7,4 @@ namespace rf
     struct NwAddr;
 }
 
-void ProcessPfPacket(const void* data, size_t len, const rf::NwAddr& addr, const rf::Player* player);
+void ProcessPfPacket(const void* data, int len, const rf::NwAddr& addr, const rf::Player* player);
